nvethernet: T26x EQOS changes
1) Added T26X EQOS bring up changes
2) Fixes added during bring up

Bug 4639097

Reviewed-on: https://git-master.nvidia.com/r/c/linux-nv-oot/+/3152935
GVS: buildbot_gerritrpt <buildbot_gerritrpt@nvidia.com>
Tested-by: Seema Khowala <seemaj@nvidia.com>

+++ drivers/net/ethernet/nvidia/nvethernet/osd.c
@@ -89,7 +89,7 @@ empty:
  *
  * @param[in] usec: Delay number in micro seconds.
  */
-static void osd_udelay(unsigned long usec)
+static void osd_udelay(unsigned long long usec)
 {
 	udelay(usec);
 }
@@ -102,7 +102,7 @@ static void osd_udelay(unsigned long usec)
  * @param[in] umin: Minimum sleep required in micro seconds.
  * @param[in] umax: Maximum sleep required in micro seconds.
  */
-static void osd_usleep_range(unsigned long umin, unsigned long umax)
+static void osd_usleep_range(unsigned long long umin, unsigned long long umax)
 {
 	usleep_range(umin, umax);
 }
